move phyical location decloration
